// ui/src/generated/api/model/dtoContestGrantResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Vertex OJ API
 * Public application endpoints and the authenticated Judge worker protocol.
 * OpenAPI spec version: 1.0
 */
import type { DtoContestGrantResponseRole } from "./dtoContestGrantResponseRole";

export interface DtoContestGrantResponse {
  groupId?: string;
  groupName?: string;
  id: number;
  role: DtoContestGrantResponseRole;
  userId?: string;
  username?: string;
}